backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.9.3
aiosignal==1.3.1
anaconda-client==1.11.2
anaconda-navigator==2.4.0
anaconda-project @ file:///opt/conda/conda-bld/anaconda-project_1660339890420/work
annotated-types==0.6.0
anyio==4.9.0
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
backports.functools-lru-cache @ file:///tmp/build/80754af9/backports.functools_lru_cache_1618170165463/work
backports.tempfile @ file:///home/linux1/recipes/ci/backports.tempfile_1610991236607/work
backports.weakref==1.0.post1
beautifulsoup4 @ file:///croot/beautifulsoup4-split_1681493039619/work
boltons @ file:///croot/boltons_1677628692245/work
bootstrap4==0.1.0
boto==2.49.0
boto3==1.26.140
botocore==1.29.140
Brotli @ file:///tmp/abs_ecyw11_7ze/croots/recipe/brotli-split_1659616059936/work
build==1.0.3
CacheControl==0.13.1
certifi==2024.2.2
cffi @ file:///croot/cffi_1700254295673/work
channels==4.0.0
channels-redis==4.1.0
chardet @ file:///home/builder/ci_310/chardet_1640804867535/work
charset-normalizer==3.3.2
cleo==2.1.0
click @ file:///croot/click_1698129812380/work
clyent==1.2.2
conda==23.3.1
conda-build==3.24.0
conda-content-trust @ file:///tmp/abs_5952f1c8-355c-4855-ad2e-538535021ba5h26t22e5/croots/recipe/conda-content-trust_1658126371814/work
conda-pack @ file:///tmp/build/80754af9/conda-pack_1611163042455/work
conda-package-handling @ file:///croot/conda-package-handling_1672865015732/work
conda-repo-cli==1.0.75
conda-token @ file:///Users/paulyim/miniconda3/envs/c3i/conda-bld/conda-token_1662660369760/work
conda-verify==3.4.2
conda_package_streaming @ file:///croot/conda-package-streaming_1670508151586/work
constantly==23.10.4
crashtest==0.4.1
cryptography==41.0.5
cssselect==1.2.0
daphne==4.0.0
dataclasses-json==0.6.7
decorator==5.1.1
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
diff-match-patch==20241021
distlib==0.3.8
distro==1.9.0
dj-database-url==2.3.0
Django==4.2.3
django-anymail==9.1
django-bootstrap-v5==1.0.11
django-bootstrap4==22.2
django-ckeditor==6.7.0
django-ckeditor-5==0.2.10
django-cors-headers==3.14.0
django-dotenv==1.4.2
django-environ==0.9.0
django-filter==25.1
django-import-export==3.2.0
django-jazzmin==3.0.1
django-js-asset==3.0.1
django-redis==5.4.0
django-rest-framework==0.1.0
django-storages==1.13.2
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
dulwich==0.21.7
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==2.0.1
Faker==18.10.1
fastapi==0.108.0
fastjsonschema==2.19.1
filelock @ file:///croot/filelock_1700591183607/work
frozenlist==1.4.1
future @ file:///croot/future_1677599870788/work
glob2 @ file:///home/linux1/recipes/ci/glob2_1610991677669/work
greenlet==3.2.2
gunicorn==21.2.0
h11==0.16.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
humanize==4.6.0
hyperframe==6.0.1
hyperlink==21.0.0
idna @ file:///croot/idna_1666125576474/work
importlib-metadata==7.0.1
incremental==22.10.0
inflection==0.5.1
installer==0.7.0
ipdb==0.13.13
ipython==8.17.2
itemadapter==0.8.0
itemloaders==1.1.0
jaraco.classes==3.3.0
jedi==0.19.1
Jinja2 @ file:///croot/jinja2_1666908132255/work
jiter==0.10.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.1
jsonschema @ file:///croot/jsonschema_1699041609003/work
jsonschema-specifications @ file:///croot/jsonschema-specifications_1699032386549/work
jupyter_core @ file:///croot/jupyter_core_1698937308754/work
keyring==24.3.0
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.62
langchain-text-splitters==0.3.8
langsmith==0.3.43
libarchive-c @ file:///tmp/build/80754af9/python-libarchive-c_1617780486945/work
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
marshmallow==3.26.1
matplot==0.1.9
matplotlib-inline==0.2.1
maxminddb==2.5.2
mdurl==0.1.2
minio==7.1.15
more-itertools==10.2.0
msgpack==1.0.7
multidict==6.0.5
mypy_extensions==1.1.0
navigator-updater @ file:///croot/navigator-updater_1695210188412/work
nbformat @ file:///croot/nbformat_1694616755618/work
numpy==1.26.4
odfpy==1.4.1
openai==1.82.0
openpyxl==3.1.2
orjson==3.10.18
packaging==23.2
pandas==2.2.0
parsel==1.8.1
parso==0.8.3
pathlib @ file:///Users/ktietz/demo/mc3/conda-bld/pathlib_1629713961906/work
pexpect==4.8.0
pgvector==0.4.2
Pillow==10.0.0
pkginfo @ file:///croot/pkginfo_1679431160147/work
platformdirs @ file:///croot/platformdirs_1692205439124/work
pluggy @ file:///tmp/build/80754af9/pluggy_1648024709248/work
ply==3.11
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
priority==1.3.0
prompt-toolkit==3.0.40
propcache==0.3.1
Protego==0.2.1
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
psycopg2==2.9.6
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycosat @ file:///croot/pycosat_1696536503704/work
pycparser==2.21
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
PyDispatcher==2.0.7
Pygments==2.16.1
PyJWT==2.6.0
pykml==0.2.0
pyloco==0.0.139
PyMuPDF==1.26.0
pyOpenSSL @ file:///croot/pyopenssl_1690223430423/work
pyproject_hooks==1.0.0
PyQt5==5.15.10
PyQt5-sip @ file:///croot/pyqt-split_1698769088074/work/pyqt_sip
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-crontab==3.3.0
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-decouple==3.8
python-dotenv==1.0.0
python-ipware==2.0.2
pytz==2024.1
PyYAML @ file:///croot/pyyaml_1698096049011/work
QtPy @ file:///croot/qtpy_1700144840038/work
queuelib==1.6.2
rapidfuzz==3.6.1
readme-renderer==40.0
redis==5.0.1
referencing @ file:///croot/referencing_1699012038513/work
requests @ file:///croot/requests_1690400202158/work
requests-file==1.5.1
requests-toolbelt @ file:///croot/requests-toolbelt_1690874004362/work
rfc3986==2.0.0
rich==13.4.2
rpds-py @ file:///croot/rpds-py_1698945930462/work
ruamel-yaml-conda @ file:///croot/ruamel_yaml_1667489728852/work
ruamel.yaml @ file:///croot/ruamel.yaml_1666304550667/work
ruamel.yaml.clib @ file:///croot/ruamel.yaml.clib_1666302247304/work
s3transfer==0.6.1
Scrapy==2.9.0
service-identity==23.1.0
shapely==2.1.0
shellingham==1.5.4
shortuuid==1.0.13
simplekml==1.3.6
SimpleWebSocketServer==0.1.2
sip @ file:///croot/sip_1698675935381/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.3.1
soupsieve @ file:///croot/soupsieve_1696347547217/work
SQLAlchemy==2.0.41
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.32.0.post1
tablib==3.7.0
tenacity==9.1.2
text-unidecode==1.3
tldextract==3.4.4
tomli @ file:///opt/conda/conda-bld/tomli_1657175507142/work
tomlkit==0.12.3
toolz @ file:///croot/toolz_1667464077321/work
tornado==6.2
tqdm @ file:///croot/tqdm_1679561862951/work
traitlets==5.13.0
trove-classifiers==2024.1.8
twine==4.0.2
Twisted==24.3.0
txaio==23.1.1
typing==3.7.4.3
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.7.1
tzdata==2024.1
ujson @ file:///opt/conda/conda-bld/ujson_1657544923770/work
uritemplate==4.1.1
urllib3==2.2.1
ushlex==0.99.1
validate-docbr==1.10.0
vine==5.1.0
virtualenv==20.25.0
w3lib==2.1.1
wcwidth==0.2.9
whitenoise==6.7.0
xlrd==2.0.1
xlwt==1.3.0
yarl==1.9.4
zope.interface==6.1
zstandard==0.23.0
